Update NI-Container.md
I filled in a few pieces of your missing info, but have almost all of this worked out in https://github.com/monomadic/ni-file

Given some of the info overlaps with older names and descriptions I was working on, I assume you were basing some of your info on my discoveries - we should get in contact if you want to work together on this. There's a lot of info I haven't documented.